+++ CHANGELOG.md
@@ -6,7 +6,7 @@ We make updates from time to time, to fix things and add functionality - we'll k
 
 ### Changed
 
-- CA Events now respect the given merge strategy when filtering by Site only. Applying no filters will continue to respect thge merge strategy. Filtering by search query string, category or (upcoming) Event Labels will continue to show all filtered models.
+- CA Events now respect the given merge strategy when filtering by Site. Filtering by search query string, category or (upcoming) Event Labels will continue to show unmerged models.
 
 ## [5.3.1] - 2026-09-10